backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-multipart
argon2-cffi
